Modified the implementation of 'BookService' Class
Renamed updateBookStatus() to updateBook() and fixed this method by actually updating the 'Book' in the List, not just updating the local Variable. \nThe searchBook() method now returns an empty List instead of 'null'. \nUpdated the addBook() method by checking if a book with the same ISBN already exists to avoid duplicates.